Update install scripts:
- cleanup
- add vim plugin installation

+++ linux/ubuntu/install_unprivileged.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+export SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+
 # TODO
 # Install python tools
 # -------------------------
@@ -14,9 +16,20 @@ eval "$(starship init bash)"
 
 # Install bash_it 
 # -------------------------
-rm -rf ~/.bash_it
+if [[ -d "${HOME}/.bash_it" ]]
+then
+    rm -rf ~/.bash_it
+fi
 git clone --depth=1 https://github.com/Bash-it/bash-it.git ~/.bash_it
-~/.bash_it/install.sh
+# -s - silent installation
+# -n - do not modify config
+# -f - overwrite existing backup
+~/.bash_it/install.sh -s -n -f
+
+
+# Install additional components for vim/neovim 
+# -------------------------
+"${SCRIPT_DIR}/vim/install.sh"
 
 # Cleanup 
 # -------------------------
